+++ R/dreamletProcessedData.R
@@ -6,22 +6,7 @@
 #' @name dreamletProcessedData-class
 #' @rdname dreamletProcessedData-class
 #' @exportClass dreamletProcessedData
-setClass("dreamletProcessedData", contains="list", slots = c(data = 'data.frame', metadata='data.frame', pkeys="vector"))
-
-# Convert SingleCellExperiment to dreamletProcessedData
-# @export
-# setAs("SingleCellExperiment", "dreamletProcessedData", function(from){
-
-# 	assayLists = lapply(assayNames(from), function(k){
-# 		obj = new("EList", list(E=assay(from, k)))
-# 		obj$isCounts = FALSE
-# 		obj
-# 	})
-# 	names(assayLists) = assayNames(from)
-
-# 	new("dreamletProcessedData", assayLists, data = as.data.frame(colData(from)), metadata = data.frame(), pkeys=vector())
-# })
-
+setClass("dreamletProcessedData", contains="list", slots = c(data = 'data.frame', metadata='data.frame', by="vector"))
 
 #' Subset with brackets
 #'
@@ -39,8 +24,8 @@ setMethod("[", signature(x="dreamletProcessedData"),
 	function(x, i){   
 		res = new("dreamletProcessedData", x@.Data[i], 
 			data = x@data, 
-			metadata = x@metadata,
-			pkeys = x@pkeys)
+			metadata = x@metadata, 
+			by = x@by)
 		names(res) = names(x)[i]
 		res
 	}