+++ R/fitVarPart.R
@@ -51,12 +51,13 @@ setGeneric("fitVarPart",
 
 
 # local definition so methods in this file have this class
-setClass("dreamletProcessedData", contains="list", slots = c(data = 'data.frame', metadata='data.frame', pkeys="vector"))
+# setClass("dreamletProcessedData", contains="list", slots = c(data = 'data.frame', metadata='data.frame', by="vector"))
 
 #' @importFrom variancePartition fitExtractVarPartModel
 #' @importFrom SummarizedExperiment colData assays
 #' @importFrom S4Vectors DataFrame as.data.frame
 #' @importFrom gtools smartbind
+#' @importFrom dplyr filter
 #' @export
 #' @rdname fitVarPart
 #' @aliases fitVarPart,dreamletProcessedData-method
@@ -87,8 +88,6 @@ setMethod("fitVarPart", "dreamletProcessedData",
 	    data_constant = droplevels(data_constant[-idx,,drop=FALSE])
 	}
 
-	pkeys = x@pkeys
-
 	# for each assay
 	resList = lapply( assays, function( k ){
 
@@ -102,9 +101,9 @@ setMethod("fitVarPart", "dreamletProcessedData",
 		ids = intersect(colnames(geneExpr), rownames(data_constant)) 
 		geneExpr = geneExpr[,ids,drop=FALSE]
 		
-		# merge data_constant and pmetadata based on pkeys and assay k
-		data2 = merge_metadata(data_constant[ids,,drop=FALSE], metadata(x), pkeys, k)
-		data2 = droplevels(data2)
+		# merge data_constant (data constant for all cell types)
+		# with metadata(sceObj)$aggr_means (data that varies)
+		data2 = merge_metadata(data_constant[ids,,drop=FALSE], metadata(x), k, x@by)
 
 		# drop any constant terms from the formula
 		form_mod = removeConstantTerms(formula, data2)
@@ -133,7 +132,7 @@ setMethod("fitVarPart", "dreamletProcessedData",
 	# Convert results to DataFrame in vpDF
 	vplst = lapply( names(resList), function(id){
 
-		# get variation partition results
+		# get variance partitioning results
 		df = resList[[id]]$df
 
 		if( nrow(df) > 0){